youtube-pipeline-project/streaming/kafka_producer/requirements.txt
kafka-python
requests
pandas
numpy